Cleanup to a class based architecture - no api change
