+++ merqube_client_lib/templates/equity_baskets/bin/get_eb_template.py
