+++ merqube_client_lib/templates/equity_baskets/bin/update_portfolio.py
