--- merqube_client_lib/templates/equity_baskets/multiple_equity_basket/modify.py
@@ -1,1 +1,0 @@
-# TODO: https://github.com/merqube/merqube-client-lib/issues/28
