+++ pylintrc
@@ -13,7 +13,8 @@ disable=C0301,  #line too long
         C0103,  #constants uppercase
         import-error, # unless youre in the venv, it won't find many packages
         unknown-option-value,  # type generation induces these
-        logging-fstring-interpolation
+        logging-fstring-interpolation,
+        too-few-public-methods  # ???
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
