+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "merqube-client-lib"
-version = "0.16.7"
+version = "0.17.0"
 description = "MerQube IndexAPI + SecAPI client library"
 authors = ["Merqube"]
 readme = "README.md"
@@ -23,9 +23,9 @@ pandas_market_calendars = "*"
 requests = "*"
 
 [tool.poetry.scripts]
-create = "merqube_client_lib.templates.equity_baskets.create_index:main"
-update_portfolio = "merqube_client_lib.templates.equity_baskets.multiple_equity_basket.update_portfolio:main"
-get_eb_template = "merqube_client_lib.templates.equity_baskets.get_eb_template:main"
+create = "merqube_client_lib.templates.equity_baskets.bin.create_index:main"
+update_portfolio = "merqube_client_lib.templates.equity_baskets.in.update_portfolio:main"
+get_eb_template = "merqube_client_lib.templates.equity_baskets.bin.get_eb_template:main"
 
 [tool.poetry.dev-dependencies]
 coverage =  {version="*"}
