+++ tests/unit/templates/test_create_index.py
@@ -1,28 +1,16 @@
 import json
 import os
 import tempfile
-from copy import deepcopy
-from unittest.mock import ANY, MagicMock, call
+from typing import Any
+from unittest.mock import MagicMock, call
 
 import pytest
 from click.testing import CliRunner
 
-from merqube_client_lib.pydantic_types import (
-    IdentifierUUIDPost,
-    IndexDefinitionPost,
-    Provider,
-    RunState,
-)
+from merqube_client_lib.pydantic_types import RunState
 from merqube_client_lib.pydantic_types import RunStateStatus as RSS
-from merqube_client_lib.templates.equity_baskets import util
-from merqube_client_lib.templates.equity_baskets.create_index import main
-from merqube_client_lib.templates.equity_baskets.multiple_equity_basket.base import (
-    inline_to_tp,
-)
-from merqube_client_lib.templates.equity_baskets.multiple_equity_basket.multieb_create import (
-    create_index,
-)
-from merqube_client_lib.templates.equity_baskets.schema import ClientIndexConfigBase
+from merqube_client_lib.templates.equity_baskets.bin.create_index import main
+from merqube_client_lib.templates.equity_baskets.util import EquityBasketIndexCreator
 
 
 @pytest.mark.parametrize("prod_run", [True, False])
@@ -32,9 +20,18 @@ def test_click_cli(prod_run, poll, monkeypatch):
     mult = MagicMock()
     ss = MagicMock()
 
-    monkeypatch.setattr("merqube_client_lib.templates.equity_baskets.create_index.dec_index", dec)
-    monkeypatch.setattr("merqube_client_lib.templates.equity_baskets.create_index.mult_index", mult)
-    monkeypatch.setattr("merqube_client_lib.templates.equity_baskets.create_index.ss_index", ss)
+    class DecCl(EquityBasketIndexCreator):
+        create = dec
+
+    class MultCl(EquityBasketIndexCreator):
+        create = mult
+
+    class SsCl(EquityBasketIndexCreator):
+        create = ss
+
+    monkeypatch.setattr("merqube_client_lib.templates.equity_baskets.bin.create_index.DC", DecCl)
+    monkeypatch.setattr("merqube_client_lib.templates.equity_baskets.bin.create_index.MEB", MultCl)
+    monkeypatch.setattr("merqube_client_lib.templates.equity_baskets.bin.create_index.SSTR", SsCl)
 
     cli = CliRunner()
 
@@ -102,144 +99,6 @@ def test_cli_fail():
         assert result.exit_code == 1
 
 
-iinfo = {
-    "base_value": 100,
-    "namespace": "testns",
-    "name": "testname",
-    "title": "testtitle",
-    "base_date": "2022-01-01",
-    "description": "testdesc",
-    "run_hour": 16,
-    "run_minute": 0,
-}
-
-iinfo = ClientIndexConfigBase.parse_obj(iinfo)
-
-
-def _get_client():
-    ident = MagicMock()
-    index = MagicMock()
-    target = MagicMock()
-
-    class FakeClient:
-        create_identifier = ident
-        create_index = index
-        replace_target_portfolio = target
-
-    return FakeClient(), ident, index, target
-
-
-def test_prod_run(v1_multi):
-    manifest = deepcopy(v1_multi)
-    del manifest["id"]
-    del manifest["status"]
-    ind = IndexDefinitionPost.parse_obj(manifest)
-
-    client, ident, index, target = _get_client()
-    create_index(
-        client=client,
-        template=ind,
-        index_info=iinfo,
-        inner_spec={},
-        prod_run=True,
-    )
-
-    assert ident.call_args_list == []
-    assert index.call_args_list == [call(index_def=ANY)]
-    assert target.call_args_list == []
-
-
-def test_prod_run_ticker(v1_multi):
-    manifest = deepcopy(v1_multi)
-    del manifest["id"]
-    del manifest["status"]
-    ind = IndexDefinitionPost.parse_obj(manifest)
-
-    inf = deepcopy(iinfo)
-    inf.bbg_ticker = "TEST"
-
-    client, ident, index, target = _get_client()
-    create_index(
-        client=client,
-        template=ind,
-        index_info=inf,
-        inner_spec={},
-        prod_run=True,
-    )
-
-    assert ident.call_args_list == [
-        call(
-            provider=Provider.bloomberg,
-            identifier_post=IdentifierUUIDPost(index_name="testname", name="TEST", namespace="testns", ticker="TEST"),
-        )
-    ]
-    assert index.call_args_list == [call(index_def=ANY)]
-    assert target.call_args_list == []
-
-
-def test_prod_run_ticker_tp(v1_multi):
-    manifest = deepcopy(v1_multi)
-    del manifest["id"]
-    del manifest["status"]
-    ind = IndexDefinitionPost.parse_obj(manifest)
-    ind.spec.index_class_args["spec"]["portfolios"]["constituents"] = [
-        {"date": "2023-06-12", "identifier": "AA.N", "quantity": 1, "security_type": "EQUITY"},
-        {"date": "2023-06-12", "identifier": "AAPL.OQ", "quantity": 2, "security_type": "EQUITY"},
-        {"date": "2023-06-13", "identifier": "AA.N", "quantity": 2, "security_type": "EQUITY"},
-        {"date": "2023-06-13", "identifier": "USD", "quantity": 100000000, "security_type": "CASH"},
-    ]
-
-    inf = deepcopy(iinfo)
-    inf.bbg_ticker = "TEST"
-
-    client, ident, index, target = _get_client()
-    create_index(
-        client=client,
-        template=ind,
-        index_info=inf,
-        inner_spec={},
-        initial_target_portfolios=inline_to_tp(ind.spec.index_class_args["spec"]["portfolios"]),
-        prod_run=True,
-    )
-
-    assert ident.call_args_list == [
-        call(
-            provider=Provider.bloomberg,
-            identifier_post=IdentifierUUIDPost(index_name="testname", name="TEST", namespace="testns", ticker="TEST"),
-        )
-    ]
-    assert index.call_args_list == [call(index_def=ANY)]
-    assert target.call_args_list == [
-        call(
-            index_id=ANY,
-            target_portfolio={
-                "positions": [
-                    {"amount": 1.0, "asset_type": "EQUITY", "identifier": "AA.N", "identifier_type": "RIC"},
-                    {"amount": 2.0, "asset_type": "EQUITY", "identifier": "AAPL.OQ", "identifier_type": "RIC"},
-                ],
-                "timestamp": "2023-06-12T00:00:00",
-                "unit_of_measure": "SHARES",
-            },
-        ),
-        call(
-            index_id=ANY,
-            target_portfolio={
-                "positions": [
-                    {"amount": 2.0, "asset_type": "EQUITY", "identifier": "AA.N", "identifier_type": "RIC"},
-                    {
-                        "amount": 100000000.0,
-                        "asset_type": "CASH",
-                        "identifier": "USD",
-                        "identifier_type": "CURRENCY_CODE",
-                    },
-                ],
-                "timestamp": "2023-06-13T00:00:00",
-                "unit_of_measure": "SHARES",
-            },
-        ),
-    ]
-
-
 @pytest.mark.parametrize(
     "seq",
     [
@@ -258,6 +117,10 @@ def test_poll(seq, monkeypatch):
     monkeypatch.setattr("time.sleep", lambda x: None)
     count = 0
 
+    class Cl(EquityBasketIndexCreator):
+        def create(self, config: dict[str, Any], prod_run: bool, poll: int):
+            pass
+
     class FakeClient:
         def get_last_index_run_state(self, index_id: str) -> RunState:
             nonlocal count
@@ -269,7 +132,7 @@ def test_poll(seq, monkeypatch):
             return RunState(status=res, error=None)
 
     if seq[-1] == RSS.SUCCEEDED:
-        util._poll(client=FakeClient(), new_id="asdf", poll=5)
+        Cl()._poll(client=FakeClient(), new_id="asdf", poll=5)
     else:
         with pytest.raises(RuntimeError):
-            util._poll(client=FakeClient(), new_id="asdf", poll=5)
+            Cl()._poll(client=FakeClient(), new_id="asdf", poll=5)
