+++ tests/unit/templates/test_multi_eb.py
@@ -1,13 +1,20 @@
 import os
 from copy import deepcopy
+from unittest.mock import ANY, MagicMock, call
 
 import pandas as pd
 import pytest
 from freezegun import freeze_time
 
-from merqube_client_lib.templates.equity_baskets.multiple_equity_basket import (
-    multieb_create as create,
+from merqube_client_lib.pydantic_types import (
+    IdentifierUUIDPost,
+    IndexDefinitionPost,
+    Provider,
 )
+from merqube_client_lib.templates.equity_baskets.multieb_create import (
+    MultiEquityBasketIndexCreator as MEB,
+)
+from merqube_client_lib.templates.equity_baskets.schema import ClientIndexConfigBase
 from tests.conftest import mock_secapi
 
 from .helpers import eb_test, eb_test_bad
@@ -164,7 +171,7 @@ def test_multi(intraday, bbg_ticker, corax_conf, has_corax, v1_multi, monkeypatc
     )
 
     eb_test(
-        func=create.create,
+        cls=MEB,
         config=good_config,
         bbg_ticker=bbg_ticker,
         expected=expected,
@@ -191,4 +198,142 @@ bad_1 = {
 
 @pytest.mark.parametrize("case", [bad_1])
 def test_multi_bad(case, v1_multi, monkeypatch):
-    eb_test_bad(func=create.create, config=case, monkeypatch=monkeypatch, template=v1_multi)
+    eb_test_bad(cls=MEB, config=case, monkeypatch=monkeypatch, template=v1_multi)
+
+
+iinfo = {
+    "base_value": 100,
+    "namespace": "testns",
+    "name": "testname",
+    "title": "testtitle",
+    "base_date": "2022-01-01",
+    "description": "testdesc",
+    "run_hour": 16,
+    "run_minute": 0,
+}
+
+iinfo = ClientIndexConfigBase.parse_obj(iinfo)
+
+
+def _get_client():
+    ident = MagicMock()
+    index = MagicMock()
+    target = MagicMock()
+
+    class FakeClient:
+        create_identifier = ident
+        create_index = index
+        replace_target_portfolio = target
+
+    return FakeClient(), ident, index, target
+
+
+def test_prod_run(v1_multi):
+    manifest = deepcopy(v1_multi)
+    del manifest["id"]
+    del manifest["status"]
+    ind = IndexDefinitionPost.parse_obj(manifest)
+
+    client, ident, index, target = _get_client()
+    MEB()._create_index(
+        client=client,
+        template=ind,
+        index_info=iinfo,
+        inner_spec={},
+        prod_run=True,
+    )
+
+    assert ident.call_args_list == []
+    assert index.call_args_list == [call(index_def=ANY)]
+    assert target.call_args_list == []
+
+
+def test_prod_run_ticker(v1_multi):
+    manifest = deepcopy(v1_multi)
+    del manifest["id"]
+    del manifest["status"]
+    ind = IndexDefinitionPost.parse_obj(manifest)
+
+    inf = deepcopy(iinfo)
+    inf.bbg_ticker = "TEST"
+
+    client, ident, index, target = _get_client()
+    MEB()._create_index(
+        client=client,
+        template=ind,
+        index_info=inf,
+        inner_spec={},
+        prod_run=True,
+    )
+
+    assert ident.call_args_list == [
+        call(
+            provider=Provider.bloomberg,
+            identifier_post=IdentifierUUIDPost(index_name="testname", name="TEST", namespace="testns", ticker="TEST"),
+        )
+    ]
+    assert index.call_args_list == [call(index_def=ANY)]
+    assert target.call_args_list == []
+
+
+def test_prod_run_ticker_tp(v1_multi):
+    manifest = deepcopy(v1_multi)
+    del manifest["id"]
+    del manifest["status"]
+    ind = IndexDefinitionPost.parse_obj(manifest)
+    ind.spec.index_class_args["spec"]["portfolios"]["constituents"] = [
+        {"date": "2023-06-12", "identifier": "AA.N", "quantity": 1, "security_type": "EQUITY"},
+        {"date": "2023-06-12", "identifier": "AAPL.OQ", "quantity": 2, "security_type": "EQUITY"},
+        {"date": "2023-06-13", "identifier": "AA.N", "quantity": 2, "security_type": "EQUITY"},
+        {"date": "2023-06-13", "identifier": "USD", "quantity": 100000000, "security_type": "CASH"},
+    ]
+
+    inf = deepcopy(iinfo)
+    inf.bbg_ticker = "TEST"
+
+    client, ident, index, target = _get_client()
+    MEB()._create_index(
+        client=client,
+        template=ind,
+        index_info=inf,
+        inner_spec={},
+        initial_target_portfolios=MEB()._inline_to_tp(ind.spec.index_class_args["spec"]["portfolios"]),
+        prod_run=True,
+    )
+
+    assert ident.call_args_list == [
+        call(
+            provider=Provider.bloomberg,
+            identifier_post=IdentifierUUIDPost(index_name="testname", name="TEST", namespace="testns", ticker="TEST"),
+        )
+    ]
+    assert index.call_args_list == [call(index_def=ANY)]
+    assert target.call_args_list == [
+        call(
+            index_id=ANY,
+            target_portfolio={
+                "positions": [
+                    {"amount": 1.0, "asset_type": "EQUITY", "identifier": "AA.N", "identifier_type": "RIC"},
+                    {"amount": 2.0, "asset_type": "EQUITY", "identifier": "AAPL.OQ", "identifier_type": "RIC"},
+                ],
+                "timestamp": "2023-06-12T00:00:00",
+                "unit_of_measure": "SHARES",
+            },
+        ),
+        call(
+            index_id=ANY,
+            target_portfolio={
+                "positions": [
+                    {"amount": 2.0, "asset_type": "EQUITY", "identifier": "AA.N", "identifier_type": "RIC"},
+                    {
+                        "amount": 100000000.0,
+                        "asset_type": "CASH",
+                        "identifier": "USD",
+                        "identifier_type": "CURRENCY_CODE",
+                    },
+                ],
+                "timestamp": "2023-06-13T00:00:00",
+                "unit_of_measure": "SHARES",
+            },
+        ),
+    ]
